Move testnet generate to network generate to simplify number of subcommands
