+++ cmd/testnet/testnet.go
@@ -8,8 +8,9 @@ import (
 
 // testnetCmd represents the config command
 var testnetCmd = &cobra.Command{
-	Use:   "testnet",
-	Short: "Utilities for working with chia testnets",
+	Use:        "testnet",
+	Short:      "Utilities for working with chia testnets",
+	Deprecated: "\nThe testnet subcommand is deprecated. Please use the 'network' subcommand instead\n",
 }
 
 func init() {
